Add Matrix server reachability check for project list items
